+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 * Enable `aabb` feature by default
+* Ability to offset an `Aabb` shape
 
 
 ## [0.2.1] - 2024-09-01